// client/src/models/items/weapons/weapon.types.ts
export enum WeaponType {
  BRONZE_SWORD = 'bronze sword',
  BRONZE_AXE = 'bronze axe',
  BRONZE_BOW = 'bronze bow',
  BRONZE_SHIELD = 'bronze shield',
  BRONZE_SPEAR = 'bronze spear',

  IRON_SWORD = 'iron sword',
  IRON_AXE = 'iron axe',
  IRON_BOW = 'iron bow',
  IRON_SHIELD = 'iron shield',
  IRON_SPEAR = 'iron spear',

  GOLD_SWORD = 'gold sword',
  GOLD_AXE = 'gold axe',
  GOLD_BOW = 'gold bow',
  GOLD_SHIELD = 'gold shield',
  GOLD_SPEAR = 'gold spear',

  TURQUOISE_SWORD = 'turqouise sword',
  TURQUOISE_AXE = 'turqouise axe',
  TURQUOISE_BOW = 'turqouise bow',
  TURQUOISE_SHIELD = 'turqouise shield',
  TURQUOISE_SPEAR = 'turqouise spear',

  ZAPHIRE_SWORD = 'zaphire sword',
  ZAPHIRE_AXE = 'zaphire axe',
  ZAPHIRE_BOW = 'zaphire bow',
  ZAPHIRE_SHIELD = 'zaphire shield',
  ZAPHIRE_SPEAR = 'zaphire spear',

  EMERALD_SWORD = 'emerald sword',
  EMERALD_AXE = 'emerald axe',
  EMERALD_BOW = 'emerald bow',
  EMERALD_SHIELD = 'emerald shield',
  EMERALD_SPEAR = 'emerald spear',

  RUBI_SWORD = 'rubi sword',
  RUBI_AXE = 'rubi axe',
  RUBI_BOW = 'rubi bow',
  RUBI_SHIELD = 'rubi shield',
  RUBI_SPEAR = 'rubi spear',
}